requirements.txt
asgiref==3.5.2
Django==4.0.7
djangorestframework==3.13.1
drf-extensions==0.7.1
et-xmlfile==1.1.0
openpyxl==3.0.10
psycopg2==2.9.3
pytz==2022.2.1
sqlparse==0.4.2
